+++ packages/destination-actions/src/destinations/engage/sendgrid/sendEmail/Region.ts
@@ -1,0 +1,1 @@
+export type Region = 'us-west-2' | 'eu-west-1'
