+++ packages/destination-actions/src/destinations/engage/sendgrid/sendEmail/UnlayerResponse.ts
@@ -1,0 +1,12 @@
+export interface UnlayerResponse {
+  success: boolean
+  data: {
+    html: string
+    chunks: {
+      css: string
+      js: string
+      fonts: string[]
+      body: string
+    }
+  }
+}
