+++ packages/destination-actions/src/destinations/engage/sendgrid/sendEmail/actionDefinition.ts
@@ -1,414 +1,8 @@
-import { ActionDefinition, IntegrationError, RequestClient, RequestOptions } from '@segment/actions-core'
+import { ActionDefinition, IntegrationError } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
-import { Liquid as LiquidJs } from 'liquidjs'
-import cheerio from 'cheerio'
-import { htmlEscape } from 'escape-goat'
-import { Logger, StatsClient } from '@segment/actions-core/destination-kit'
-import { apiLookupActionFields, apiLookupLiquidKey, performApiLookups } from '../utils/api-lookups'
-import { Profile } from '../utils/types'
-
-const Liquid = new LiquidJs()
-
-type Region = 'us-west-2' | 'eu-west-1'
-
-const insertEmailPreviewText = (html: string, previewText: string): string => {
-  const $ = cheerio.load(html)
-
-  // See https://www.litmus.com/blog/the-little-known-preview-text-hack-you-may-want-to-use-in-every-email/
-  $('body').prepend(`
-    <div style='display: none; max-height: 0px; overflow: hidden;'>
-      ${htmlEscape(previewText)}
-    </div>
-
-    <div style='display: none; max-height: 0px; overflow: hidden;'>
-      ${'&nbsp;&zwnj;'.repeat(13)}&nbsp;
-    </div>
-  `)
-
-  return $.html()
-}
-
-const insertUnsubscribeLinks = (
-  html: string,
-  emailProfile: any,
-  spaceId: string,
-  statsClient: StatsClient | undefined,
-  tags: string[],
-  groupId?: string,
-  logger?: Logger | undefined
-): string => {
-  const globalUnsubscribeLink = emailProfile?.unsubscribeLink
-  const preferencesLink = emailProfile?.preferencesLink
-  const unsubscribeLinkRef = 'a[href*="[upa_unsubscribe_link]"]'
-  const preferencesLinkRef = 'a[href*="[upa_preferences_link]"]'
-  const $ = cheerio.load(html)
-  if (groupId) {
-    const group = emailProfile?.groups.find((group: { id: string }) => group?.id === groupId)
-    const groupUnsubscribeLink = group?.groupUnsubscribeLink
-    $(unsubscribeLinkRef).each(function () {
-      logger?.info(`TE Messaging: Email Group Unsubscribe link replaced  - ${spaceId} ${groupId}`)
-      statsClient?.incr('actions-personas-messaging-sendgrid.replaced_group_unsubscribe_link', 1, tags)
-      $(this).attr('href', groupUnsubscribeLink)
-    })
-  } else {
-    $(unsubscribeLinkRef).each(function () {
-      logger?.info(`TE Messaging: Email Global Unsubscribe link replaced  - ${spaceId}`)
-      statsClient?.incr('actions-personas-messaging-sendgrid.replaced_global_unsubscribe_link', 1, tags)
-      $(this).attr('href', globalUnsubscribeLink)
-    })
-  }
-  $(preferencesLinkRef).each(function () {
-    if (!preferencesLink) {
-      // Remove the Manage Preferences link placeholder and the pipe (' | ') symbol
-      $(this.parent?.children).each(function () {
-        if ($(this).text() == ' | ') {
-          $(this).remove()
-        }
-      })
-      $(this).remove()
-      logger?.info(`TE Messaging: Email Preferences link removed from the html body  - ${spaceId}`)
-      statsClient?.incr('actions-personas-messaging-sendgrid.removed_preferences_link', 1, tags)
-    } else {
-      $(this).attr('href', preferencesLink)
-      logger?.info(`TE Messaging: Email Preferences link replaced  - ${spaceId}`)
-      statsClient?.incr('actions-personas-messaging-sendgrid.replaced_preferences_link', 1, tags)
-    }
-  })
-
-  return $.html()
-}
-
-// These profile calls will be removed when Profile sync can fetch external_id
-const getProfileApiEndpoint = (environment: string, region?: Region): string => {
-  const domainName = region === 'eu-west-1' ? 'profiles.euw1.segment' : 'profiles.segment'
-  const topLevelName = environment === 'production' ? 'com' : 'build'
-  return `https://${domainName}.${topLevelName}`
-}
-
-const fetchProfileTraits = async (
-  request: RequestClient,
-  settings: Settings,
-  profileId: string,
-  statsClient: StatsClient | undefined,
-  tags: string[],
-  logger?: Logger | undefined
-): Promise<Record<string, string>> => {
-  try {
-    const endpoint = getProfileApiEndpoint(settings.profileApiEnvironment, settings.region as Region)
-    const response = await request(
-      `${endpoint}/v1/spaces/${settings.spaceId}/collections/users/profiles/user_id:${profileId}/traits?limit=200`,
-      {
-        headers: {
-          authorization: `Basic ${Buffer.from(settings.profileApiAccessToken + ':').toString('base64')}`,
-          'content-type': 'application/json'
-        }
-      }
-    )
-    tags.push(`profile_status_code:${response.status}`)
-    statsClient?.incr('actions-personas-messaging-sendgrid.profile_invoked', 1, tags)
-
-    const body = await response.json()
-    return body.traits
-  } catch (error) {
-    logger?.error(`TE Messaging: Email profile traits request failure - ${settings.spaceId} - [${error}]`)
-    tags.push('reason:profile_error')
-    statsClient?.incr('actions-personas-messaging-sendgrid.error', 1, tags)
-    throw new IntegrationError('Unable to get profile traits for the email message', 'Email trait fetch failure', 500)
-  }
-}
-
-const isRestrictedDomain = (email: string): boolean => {
-  const restricted = ['gmailx.com', 'yahoox.com', 'aolx.com', 'hotmailx.com']
-  const matches = /^.+@(.+)$/.exec(email.toLowerCase())
-
-  if (!matches) {
-    return false
-  }
-
-  const domain = matches[1]
-  return restricted.includes(domain)
-}
-
-interface UnlayerResponse {
-  success: boolean
-  data: {
-    html: string
-    chunks: {
-      css: string
-      js: string
-      fonts: string[]
-      body: string
-    }
-  }
-}
-
-const generateEmailHtml = async (
-  request: RequestClient,
-  settings: Settings,
-  design: string,
-  statsClient: StatsClient | undefined,
-  tags: string[],
-  logger?: Logger | undefined
-): Promise<string> => {
-  try {
-    statsClient?.incr('actions-personas-messaging-sendgrid.unlayer_request', 1, tags)
-    const response = await request('https://api.unlayer.com/v2/export/html', {
-      method: 'POST',
-      headers: {
-        authorization: `Basic ${Buffer.from(`${settings.unlayerApiKey}:`).toString('base64')}`,
-        'content-type': 'application/json'
-      },
-      body: JSON.stringify({
-        displayMode: 'email',
-        design: JSON.parse(design)
-      })
-    })
-
-    const body = await response.json()
-    return (body as UnlayerResponse).data.html
-  } catch (error) {
-    logger?.error(`TE Messaging: Email export request failure - ${settings.spaceId} - [${error}]`)
-    tags.push('reason:generate_email_html')
-    statsClient?.incr('actions-personas-messaging-sendgrid.error', 1, tags)
-    throw new IntegrationError('Unable to export email as HTML', 'Export HTML failure', 400)
-  }
-}
-
-const parseTemplating = async (
-  content: string,
-  liquidData: {
-    profile: Profile
-    [apiLookupLiquidKey]?: Record<string, unknown>
-  },
-  contentType: string,
-  statsClient: StatsClient | undefined,
-  tags: string[],
-  settings: Settings,
-  logger?: Logger | undefined
-) => {
-  try {
-    const parsedContent = await Liquid.parseAndRender(content, liquidData)
-    return parsedContent
-  } catch (error) {
-    logger?.error(`TE Messaging: Email templating parse failure - ${settings.spaceId} - [${error}]`)
-    tags.push('reason:parse_templating')
-    statsClient?.incr('actions-personas-messaging-sendgrid.error', 1, tags)
-    throw new IntegrationError(
-      `Unable to parse templating in email ${contentType}`,
-      `${contentType} templating parse failure`,
-      400
-    )
-  }
-}
-
-const EXTERNAL_ID_KEY = 'email'
-
-const attemptEmailDelivery = async (
-  request: RequestClient,
-  settings: Settings,
-  payload: Payload,
-  logger: Logger | undefined,
-  statsClient: StatsClient | undefined,
-  tags: string[],
-  byPassSubscription: boolean
-) => {
-  let traits
-  const emailProfile = payload?.externalIds?.find((meta) => meta.type === 'email')
-  if (payload.traitEnrichment) {
-    traits = payload?.traits ? payload?.traits : JSON.parse('{}')
-  } else {
-    if (!payload.userId) {
-      logger?.error(
-        `TE Messaging: Unable to process email, no userId provided and trait enrichment disabled - ${settings.spaceId}`
-      )
-      tags.push('reason:missing_user_id')
-      statsClient?.incr('actions-personas-messaging-sendgrid.error', 1, tags)
-      throw new IntegrationError(
-        'Unable to process email, no userId provided and trait enrichment disabled',
-        'Invalid parameters',
-        400
-      )
-    }
-    traits = await fetchProfileTraits(request, settings, payload.userId, statsClient, tags, logger)
-  }
-
-  const profile: Profile = {
-    email: emailProfile?.id,
-    traits
-  }
-
-  const toEmail = payload.toEmail || profile.email
-
-  if (!toEmail) {
-    return
-  }
-
-  if (isRestrictedDomain(toEmail)) {
-    logger?.error(
-      `TE Messaging: Emails with gmailx.com, yahoox.com, aolx.com, and hotmailx.com domains are blocked - ${settings.spaceId}`
-    )
-    tags.push('reason:restricted_domain')
-    statsClient?.incr('actions-personas-messaging-sendgrid.error', 1, tags)
-    throw new IntegrationError(
-      'Emails with gmailx.com, yahoox.com, aolx.com, and hotmailx.com domains are blocked.',
-      'Invalid input',
-      400
-    )
-  }
-
-  let name
-  if (traits.first_name && traits.last_name) {
-    name = `${traits.first_name} ${traits.last_name}`
-  } else if (traits.firstName && traits.lastName) {
-    name = `${traits.firstName} ${traits.lastName}`
-  } else if (traits.name) {
-    name = traits.name
-  } else {
-    name = traits.first_name || traits.last_name || traits.firstName || traits.lastName || 'User'
-  }
-
-  const bcc = JSON.parse(payload.bcc ?? '[]')
-  const [parsedSubject, apiLookupData] = await Promise.all([
-    parseTemplating(payload.subject, { profile }, 'Subject', statsClient, tags, settings, logger),
-    performApiLookups(request, payload.apiLookups, profile, statsClient, tags, settings, logger)
-  ])
-
-  let parsedBodyHtml
-
-  if (payload.bodyUrl && settings.unlayerApiKey) {
-    const { content: body } = await request(payload.bodyUrl, { method: 'GET', skipResponseCloning: true })
-    const bodyHtml =
-      payload.bodyType === 'html' ? body : await generateEmailHtml(request, settings, body, statsClient, tags, logger)
-    parsedBodyHtml = await parseTemplating(
-      bodyHtml,
-      { profile, [apiLookupLiquidKey]: apiLookupData },
-      'Body',
-      statsClient,
-      tags,
-      settings,
-      logger
-    )
-  } else {
-    parsedBodyHtml = await parseTemplating(
-      payload.bodyHtml ?? '',
-      { profile, [apiLookupLiquidKey]: apiLookupData },
-      'Body HTML',
-      statsClient,
-      tags,
-      settings,
-      logger
-    )
-  }
-
-  // only include preview text in design editor templates
-  if (payload.bodyType === 'design' && payload.previewText) {
-    const parsedPreviewText = await parseTemplating(
-      payload.previewText,
-      { profile },
-      'Preview text',
-      statsClient,
-      tags,
-      settings,
-      logger
-    )
-    parsedBodyHtml = insertEmailPreviewText(parsedBodyHtml, parsedPreviewText)
-  }
-
-  parsedBodyHtml = insertUnsubscribeLinks(
-    parsedBodyHtml,
-    emailProfile,
-    settings.spaceId,
-    statsClient,
-    tags,
-    payload.groupId,
-    logger
-  )
-
-  try {
-    statsClient?.incr('actions-personas-messaging-sendgrid.request', 1, tags)
-    const mailContentSubscriptionHonored = {
-      personalizations: [
-        {
-          to: [
-            {
-              email: toEmail,
-              name: name
-            }
-          ],
-          bcc: bcc.length > 0 ? bcc : undefined,
-          custom_args: {
-            ...payload.customArgs,
-            source_id: settings.sourceId,
-            space_id: settings.spaceId,
-            user_id: payload.userId ?? undefined,
-            __segment_internal_external_id_key__: EXTERNAL_ID_KEY,
-            __segment_internal_external_id_value__: profile[EXTERNAL_ID_KEY]
-          }
-        }
-      ],
-      from: {
-        email: payload.fromEmail,
-        name: payload.fromName
-      },
-      reply_to: {
-        email: payload.replyToEmail,
-        name: payload.replyToName
-      },
-      subject: parsedSubject,
-      content: [
-        {
-          type: 'text/html',
-          value: parsedBodyHtml
-        }
-      ],
-      tracking_settings: {
-        subscription_tracking: {
-          enable: true,
-          substitution_tag: '[unsubscribe]'
-        }
-      }
-    }
-    let mailContent
-    if (byPassSubscription) {
-      mailContent = {
-        ...mailContentSubscriptionHonored,
-        mail_settings: {
-          bypass_list_management: {
-            enable: true
-          }
-        }
-      }
-      statsClient?.incr('actions-personas-messaging-sendgrid.request.by_pass_subscription', 1, tags)
-    } else {
-      mailContent = mailContentSubscriptionHonored
-      statsClient?.incr('actions-personas-messaging-sendgrid.request.dont_pass_subscription', 1, tags)
-    }
-    const req: RequestOptions = {
-      method: 'post',
-      headers: {
-        authorization: `Bearer ${settings.sendGridApiKey}`
-      },
-      json: mailContent
-    }
-    statsClient?.set('actions-personas-messaging-sendgrid.request_body_size', JSON.stringify(req).length, tags)
-    const response = await request('https://api.sendgrid.com/v3/mail/send', req)
-    tags.push(`sendgrid_status_code:${response.status}`)
-    statsClient?.incr('actions-personas-messaging-sendgrid.response', 1, tags)
-    if (payload?.eventOccurredTS != undefined) {
-      statsClient?.histogram(
-        'actions-personas-messaging-sendgrid.eventDeliveryTS',
-        Date.now() - new Date(payload?.eventOccurredTS).getTime(),
-        tags
-      )
-    }
-    return response
-  } catch (error: unknown) {
-    logger?.error(`TE Messaging: Email message request failure - ${settings.spaceId} - [${error}]`)
-    statsClient?.incr('actions-personas-messaging-sendgrid.request-failure', 1, tags)
-    throw new IntegrationError('Unable to send email message', 'SendGrid API request failure', 500)
-  }
-}
+import { apiLookupActionFields } from '../utils/api-lookups'
+import { attemptEmailDelivery } from './attemptEmailDelivery'
 
 export const actionDefinition: ActionDefinition<Settings, Payload> = {
   title: 'Send Email',
